Avoid warnings about non-exhaustive pattern matching.
This change concerns only the "! MIN_VERSION_ghc_prim(0,8,0)" case, in which
this package is needed only for forward compatibility with the opposite case.

With GHC 8.8.4 and a pattern synonym such as "pattern W8# x = W.W8# x",
GHC considers the pattern (W8# _) to be non-exhaustive, unlike the actual
data constructor GHC.Word.W8#.

This change avoids that problem by reexporting GHC.Word.W8# instead of a new
pattern.  Naturally, it does the same for W16#, W32#, I8#, I16#, and I32#.